Add files via upload
Kalman Filter For Stationary Objects.
Thermal camera installed on an Airborne platform observed position of a stationary target.Thus Obtained position is noisy because Thermal camera is affecting certain vibration due to malfunctioned platform. The objective is to estimate the position (along x-axis) of target. The measurement noise is additive having Gaussian PDF with zero mean and standard deviation of 1 m. Observation time is 5 seconds and sampling time is 0.1 second.